Added Kurtosis version check

+++ start.sh
@@ -2,12 +2,19 @@
 ENCLAVE_ID="wakurtosis"
 PROMETHEUS_YML="prometheus.yml"
 
+# Check Kurtosis version
+kurtosis_version=$(kurtosis version)
+if [ "$kurtosis_version" != "0.57.4" ]; then
+    echo "Kurtosis version $kurtosis_version is not supported. Please use version 0.57.4*"
+    exit 1
+fi
+
 # TODO: Check if the enclave already exists 
 kurtosis enclave stop $ENCLAVE_ID
 kurtosis enclave rm $ENCLAVE_ID
 
 # Start the Kurtosis enclave
-kurtosis run ./main.star --enclave-id $ENCLAVE_ID
+kurtosis run main.star --enclave-id $ENCLAVE_ID
 
 # Fetch the targets from the Kurtosis output
 # TODO: Fetch the targets directly in the Starlark script
@@ -17,8 +24,10 @@ targets=$(kurtosis enclave inspect $ENCLAVE_ID | grep 'prometheus' | sed -e 's/^
 echo "Building Prometheus targets  ..."
 echo "[{\"labels\": {\"job\": \"wakurtosis\"}, \"targets\" : [$targets] } ]" | tee './targets.json' > /dev/null
 
-# Start / Restart Prometheus on http://localhost:9090
-echo "Start Prometheus with on http://localhost:9090 with: prometheus --config.file=./prometheus.yml &"
+echo "Prometheus targets are: $targets"
 
-# Starting Grafana on http://localhost:3000 with: brew services restart grafana
+# Start Prometheus on http:/localhost:9090
+prometheus --config.file=./prometheus.yml
 
+# Follow the instructions to setup Grafana on https://github.com/waku-org/nwaku/blob/master/docs/operators/how-to/monitor.md
+# Grafana server http://localhost:3000
